поставил webpack, разбил на модули
