+++ src/components/showMoreButton.js
@@ -1,0 +1,7 @@
+const createShowMoreButtonTemplate = () => {
+  return (
+    `<button class="films-list__show-more">Show more</button>`
+  );
+};
+
+export {createShowMoreButtonTemplate};
