+++ src/components/statisticContainer.js
@@ -1,0 +1,8 @@
+const createStatisticContainerTemplate = () => {
+  return (
+    `<section class="statistic">
+    </section>`
+  );
+};
+
+export {createStatisticContainerTemplate};
